Re #28987 SANS wav switch data struct. and remove bundle pre-pro
Switches the pair of lists to a list of pairs, representing the start
and end of each wav range. The data class also contains the abs max and
min value, which is used quite frequently within the codebase.

Removes the bundle pre-calc step which creates a loop per wav range
pair, in prep of moving the processing into the workflow, rather than
running the workflow n times